Tokens

+++ api/hosts/remove_host_media.js
@@ -7,32 +7,43 @@ const Grid = require('gridfs-stream');
 const Host = require("../../models/host/host");
 const Media = require("../../models/media/media");
 const mongoose = require('mongoose');
+const firebase = require("../../helpers/firebase");
+const User = require("../../models/user/user");
 
 module.exports.call = async  function (req, res) {
 
   let functionName = "remove-host-media";
-
   await paramCheck.checkParameters(req, functionName);
-  var hostWithMedia = await Host.findOneAndUpdate( { _id: req.params.hostId }, { $pull: { media: req.body.idMedia } },{ new: true });
-  if(hostWithMedia != null){
-    var removedMedia = await Media.findOneAndRemove({  _id: req.body.idMedia });
-    if(removedMedia != null){
-      console.log(removedMedia);
-      var conn = await mongooseInit.initMongoDBConnection();
-      const gfs = await Grid(conn.db, mongoose.mongo);
-      await gfs.collection('uploads');
-      gfs.files.remove({filename: removedMedia.name });
 
-      let result = {
-        status: "success",
-        message: "a media was remove from this host",
-        _id: req.params.hostId,
+  var user = await User.findOne({_id : req.payload._id});
+
+  if(user!=null && user["hosts"].indexOf(req.params.hostId)!=-1){
+    var hostWithMedia = await Host.findOneAndUpdate( { _id: req.params.hostId }, { $pull: { media: req.body.idMedia } },{ new: true });
+    if(hostWithMedia != null){
+      var removedMedia = await Media.findOneAndRemove({  _id: req.body.idMedia });
+      if(removedMedia != null){
+        console.log(removedMedia);
+        var conn = await mongooseInit.initMongoDBConnection();
+        const gfs = await Grid(conn.db, mongoose.mongo);
+        await gfs.collection('uploads');
+        gfs.files.remove({filename: removedMedia.name });
+
+        let result = {
+          status: "success",
+          message: "a media was remove from this host",
+          _id: req.params.hostId,
+        }
+        res.json(result);
+        console.log("a media was remove from this host");
       }
-      res.json(result);
-      console.log("a media was remove from this host");
-    }
-    else{
-      console.log("no media with this id were found");
+      else{
+        console.log("no media with this id were found");
+      }
     }
   }
+  else{
+    console.log("user is not admin or does not exist");
+    res.send({"error":"user is not authorise or does not exist"});
+  }
+
 };
